Add SetDatapoints to Broker-interface
This enables setting actuator targets using the Broker interface.
Sensor values, attributes and the actual value of actuators are still
set through the Collector interface.

Change the "set" method of the cli to use the Broker interface.
Add "feed" method to the cli which behaves as the previous "set" method.