Fix clippy::uninlined_format_args lint
https://rust-lang.github.io/rust-clippy/master/#uninlined_format_args

This lint was promoted from "pedantic" in Rust 1.67.0.  It was
subsequently [demoted to "pedantic"
again](https://github.com/rust-lang/rust/pull/107743) in 1.67.1, so one
solution would be to just use another version for now.

It does however seem likely that it will be reintroduced at some point,
so applying the fix.

https://github.com/rust-lang/rust-clippy/pull/10265