Add example token for actuation and provision
Intended use-case is tests with KUKSA.val Client if you want to be able to do
both setValue and setTargetValue

Also removing "width" from provide-vehicle-speed.
If needed in some form that seems more to be a use-case-specific token
which then better shall be stored within the service or test-case using it.